requirements.txt
# On macOS < Monterey (12.0, platform_release = 21.x), pyobjc 9.0 is the last version that works, #1324
beautifulsoup4
bitmath>=1.3.3.1,<1.4.0.0
bpylist2>=4.1.1,<5.0.0
cgmetadata>=0.2.0; sys_platform == 'darwin'
Click>=8.1.3,<9.0
mac-alias>=2.2.2,<3.0.0; sys_platform == 'darwin'
makelive>=0.6.2; sys_platform == 'darwin'
markdown2
Mako>=1.2.2,<1.3.0
more-itertools>=8.8.0,<9.0.0
objexplore>=1.6.3,<2.0.0
osxmetadata>=1.4.0,<2.0.0; sys_platform == 'darwin'
packaging>=21.3
pathvalidate>=2.4.1,<4.0.0
photoscript>=0.5.0; sys_platform == 'darwin'
pip
psutil
ptpython>=3.0.20,<4.0.0
pyobjc-core>=9.0,<10.0; sys_platform == 'darwin' and platform_release < '22.0'
pyobjc-core>=9.0; sys_platform == 'darwin' and platform_release >= '22.0'
pyobjc-framework-AppleScriptKit>=9.0,<10.0; sys_platform == 'darwin' and platform_release < '22.0'
pyobjc-framework-AppleScriptKit>=9.0; sys_platform == 'darwin' and platform_release >= '22.0'
pyobjc-framework-AppleScriptObjC>=9.0,<10.0; sys_platform == 'darwin' and platform_release < '22.0'
pyobjc-framework-AppleScriptObjC>=9.0; sys_platform == 'darwin' and platform_release >= '22.0'
pyobjc-framework-AVFoundation>=9.0,<10.0; sys_platform == 'darwin' and platform_release < '22.0'
pyobjc-framework-AVFoundation>=9.0; sys_platform == 'darwin' and platform_release >= '22.0'
pyobjc-framework-Cocoa>=9.0,<10.0; sys_platform == 'darwin' and platform_release < '22.0'
pyobjc-framework-Cocoa>=9.0; sys_platform == 'darwin' and platform_release >= '22.0'
pyobjc-framework-CoreServices>=9.0,<10.0; sys_platform == 'darwin' and platform_release < '22.0'
pyobjc-framework-CoreServices>=9.0; sys_platform == 'darwin' and platform_release >= '22.0'
pyobjc-framework-Metal>=9.0,<10.0; sys_platform == 'darwin' and platform_release < '22.0'
pyobjc-framework-Metal>=9.0; sys_platform == 'darwin' and platform_release >= '22.0'
pyobjc-framework-Photos>=9.0,<10.0; sys_platform == 'darwin' and platform_release < '22.0'
pyobjc-framework-Photos>=9.0; sys_platform == 'darwin' and platform_release >= '22.0'
pyobjc-framework-Quartz>=9.0,<10.0; sys_platform == 'darwin' and platform_release < '22.0'
pyobjc-framework-Quartz>=9.0; sys_platform == 'darwin' and platform_release >= '22.0'
pyobjc-framework-Vision>=9.0,<10.0; sys_platform == 'darwin' and platform_release < '22.0'
pyobjc-framework-Vision>=9.0; sys_platform == 'darwin' and platform_release >= '22.0'
pytimeparse2>=1.4.0,<2.0.0
PyYAML>=6.0.0,<7.0.0
requests>=2.27.1,<3.0.0
rich_theme_manager>=0.11.0
rich>=13.5.2,<14.0.0
shortuuid>=1.0.9,<2.0.0
strpdatetime>=0.4.0
tenacity>=8.0.1,<9.0.0
textx>=4.0.1,<5.0.0
toml>=0.10.2,<0.11.0
tzdata
utitools>=0.3.0
whenever>=0.8.3,<0.9.0
wrapt>=1.14.1,<2.0.0
wurlitzer>=3.0.2,<4.0.0
xdg-base-dirs>=6.0.0; python_version >= '3.10'
xdg==5.1.1; python_version <= '3.9'
